src/app/+dashboard/+model-viewer/model-viewer.component.spec.ts
import {
  beforeEach,
  beforeEachProviders,
  describe,
  expect,
  it,
  inject,
} from '@angular/core/testing';
import { ComponentFixture, TestComponentBuilder } from '@angular/compiler/testing';
import { Component } from '@angular/core';
import { By } from '@angular/platform-browser';
import { ModelViewerComponent } from './model-viewer.component';

describe('Component: ModelViewer', () => {
  let builder: TestComponentBuilder;

  beforeEachProviders(() => [ModelViewerComponent]);
  beforeEach(inject([TestComponentBuilder], function (tcb: TestComponentBuilder) {
    builder = tcb;
  }));

  it('should inject the component', inject([ModelViewerComponent],
      (component: ModelViewerComponent) => {
    expect(component).toBeTruthy();
  }));

  it('should create the component', inject([], () => {
    return builder.createAsync(ModelViewerComponentTestController)
      .then((fixture: ComponentFixture<any>) => {
        let query = fixture.debugElement.query(By.directive(ModelViewerComponent));
        expect(query).toBeTruthy();
        expect(query.componentInstance).toBeTruthy();
      });
  }));
});

@Component({
  selector: 'test',
  template: `
    <app-model-viewer></app-model-viewer>
  `,
  directives: [ModelViewerComponent]
})
class ModelViewerComponentTestController {
}

